Finished testing MaximizeScoreStrategy
Got rid of the ability for a move to track it's score, as it was not useful (every strategy has a different idea of what a score means).

Overrided Equals and Hashcode for Moves.

Continued to improve the general testing environemnt, adding capability for the TranscriptMockModelAdapter to track what positions were checked for easy testing.

Added FibingMockModel decorator for testing purposes (lies about leaglity and scores of moves)

Finished testing MaximizeScoreStrategy