Wrote StrategyPlayer and some tests for it.
Strategy player is probably done.

The tests are not but should be simple.

They also include tests for other Player implementations.

Also updated interfaces to what they are supposed to be. Holding of on pushing util I am sure these match up with the interfaces already on the main branch.